requirements.txt
openai==2.16.0
streamlit==1.53.1
yfinance==1.2.0
pandas==2.3.3
python-dotenv==1.2.1
requests==2.32.5
openpyxl==3.1.5
